requirements.txt
requests
beautifulsoup4
pandas
matplotlib
python-dotenv